Move and rename streaming logic into utils
